requirements.txt
altgraph==0.17
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
elevate==0.1.3
future==0.18.2
futures==3.1.1
idna==2.10
importlib-metadata==3.10.0
Jinja2==3.0.1
jsonpickle==2.0.0
jsonrpclib==0.1.7
lxml==4.6.3
macholib==1.14
markdown2==2.4.0
MarkupSafe==2.0.1
mock-server==0.3.9
packaging==20.9
pcapy==0.11.4
pefile==2021.5.13
pip==21.0.1
psutil==5.8.0
py-bcrypt==0.4
py2exe==0.10.4.0
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.10.1
PyInstaller==3.6
pyinstaller-versionfile==2.0.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
setuptools==52.0.0.post20210125
sip==4.19.13
six==1.16.0
tornado==4.3
urllib3==1.26.4
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.4.1
